refactor(settings): retire Connect tab into Extensions and Debug
Redirect /settings/connect to Extensions, move Agent access into Debug,
and keep Connect row helpers for inventory grouping.
