+++ apps/app/src/react-app/domains/settings/pages/general-view.tsx
@@ -2,7 +2,6 @@
 import {
   ArrowRight,
   ArrowUpRight,
-  Cable,
   Cloud,
   Cog,
   FolderLock,
@@ -44,7 +43,6 @@ const workspaceCards: SettingsCardDefinition[] = [
 const globalCards: SettingsCardDefinition[] = [
   { tab: "ai", icon: Sparkles, title: "AI Providers", desc: "Connect services that provide AI models." },
   { tab: "cloud-account", icon: Cloud, title: "Cloud", desc: "OpenWork Cloud account and organization." },
-  { tab: "connect", icon: Cable, titleKey: "settings.tab_connect", descKey: "settings.tab_description_connect" },
   { tab: "appearance", icon: Paintbrush, title: "Appearance", desc: "Theme, font size, and display." },
   { tab: "environment", icon: Terminal, title: "Environment", desc: "Environment variables and paths." },
   { tab: "updates", icon: RefreshCcw, title: "Updates", desc: "App version and update channel." },
