+++ apps/app/tests/settings-route.test.ts
@@ -3,11 +3,16 @@ import { describe, expect, test } from "bun:test";
 import { parseSettingsPath } from "../src/react-app/shell/settings-route";
 
 describe("settings route parsing", () => {
-  test("recognizes the Connect settings tab", () => {
-    expect(parseSettingsPath("/settings/connect")).toEqual({ tab: "connect", redirectPath: null });
+  test("redirects Connect settings into Extensions", () => {
+    expect(parseSettingsPath("/settings/connect")).toEqual({
+      tab: "extensions",
+      redirectPath: "extensions",
+      extensionsSection: "all",
+    });
     expect(parseSettingsPath("/workspace/workspace_1/settings/connect")).toEqual({
-      tab: "connect",
-      redirectPath: null,
+      tab: "extensions",
+      redirectPath: "extensions",
+      extensionsSection: "all",
     });
   });
